FormPresenter::addButton: Allow a "btn-value" and a "name" option to have multiple submit buttons in a form
When a form has multiple submit buttons, the clicked button
is identified by the "name" attribute, not the "id" attribute.
The id is not sent, but if a value and name attribute exists,
it is included in the post variables.

So far, the name was set to the id, which needs to be unique,
so it was not possible to have multiple buttons with the same
name, but different value. Also, the value option is already
reserved for the display text, so I added an option 'btn-value'
that will be inserted into the value attribute of the <input> tag.